Fix error in "rust is more than safety"
Some fun history here: for some reason, when I imported this article
into Notion, it assumed that this was an embedded table of sorts. So
then when I exported it *out* of notion, it assumed it was an image of
some kind? Idk, but here's the text from the original version:

https://archive.is/LGYIs

Fixes #3